Data Class Refactors (SubstituitionMatrixHelper)
